backend/requirements.txt
# FastAPI and web server
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
python-multipart>=0.0.21

# Database
sqlmodel>=0.0.31
alembic>=1.17.2

# Authentication and security
python-jose[cryptography]>=3.5.0
pwdlib[argon2]>=0.3.0

# Configuration
pydantic-settings>=2.12.0

# Validation
email-validator>=2.3.0

# Rate limiting
slowapi>=0.1.9

# AI and MCP
openai-agents>=0.1.0
mcp>=0.1.0
fastmcp>=0.4.1

# Testing
pytest>=9.0.2
pytest-asyncio>=1.3.0
pytest-cov>=6.0.0
httpx>=0.28.1

# PostgreSQL driver
psycopg2-binary>=2.9.9
